refactor(themes): Align themes to Brand Refresh
BREAKING CHANGE: Theme names have changed, references to old names
will cause `sku build` fail.

Themes have been renamed as follows:

* seekAnz -> seekAnzClassic
* jobsDb -> jobsDbClassic
* jobStreet -> jobsStreetClassic
* jobStreetClassic -> jobStreetRC

Other changes:

* seek is a new theme that encapsulates the Brand Refresh work, it
replaces the jobsDBRebrand theme
* seekBusiness and wireframe themes are unaffected
