+++ lib/themes/index.ts
@@ -1,7 +1,7 @@
-export { default as seekAnz } from './seekAnz';
+export { default as seekAnzClassic } from './seekAnzClassic';
 export { default as seekBusiness } from './seekBusiness';
-export { default as jobsDbRebrand } from './jobsDbRebrand';
-export { default as jobsDb } from './jobsDb';
-export { default as jobStreet } from './jobStreet';
+export { default as seek } from './seek';
+export { default as jobsDbClassic } from './jobsDbClassic';
 export { default as jobStreetClassic } from './jobStreetClassic';
+export { default as jobStreetRC } from './jobStreetRC';
 export { default as wireframe } from './wireframe';
