--- lib/themes/jobStreet/tokens.ts
@@ -1,11 +1,0 @@
-import { makeTokens } from '../baseTokens/seekAsia';
-
-const brand = '#1c3f94';
-
-export default makeTokens({
-  name: 'jobStreet',
-  bodyBackground: '#e5e5e5',
-  brand,
-  brandAccent: '#fff200',
-  formAccent: brand,
-});
