--- lib/themes/jobsDbRebrand/tokens.ts
@@ -1,5 +1,0 @@
-import { makeTokens } from '../baseTokens/seekAsiaRebrand';
-
-export default makeTokens({
-  name: 'jobsDbRebrand',
-});
