+++ lib/themes/seek/tokens.ts
@@ -1,0 +1,5 @@
+import { makeTokens } from '../baseTokens/seek';
+
+export default makeTokens({
+  name: 'seek',
+});
