+++ lib/themes/seekAnzClassic/tokens.ts
@@ -1,7 +1,7 @@
-import { makeTokens } from '../baseTokens/seekAnz';
+import { makeTokens } from '../baseTokens/seekAnzClassic';
 
 export default makeTokens({
-  name: 'seekAnz',
+  name: 'seekAnzClassic',
   brand: '#0d3880',
   brandAccent: '#e60278',
   formAccent: '#2765cf',
