--- themes/jobStreet.ts
@@ -1,1 +1,0 @@
-export { default } from '../lib/themes/jobStreet';
