+++ themes/jobStreetRC.ts
@@ -1,0 +1,1 @@
+export { default } from '../lib/themes/jobStreetRC';
