--- themes/jobsDb.ts
@@ -1,1 +1,0 @@
-export { default } from '../lib/themes/jobsDb';
