+++ themes/jobsDbClassic.ts
@@ -1,0 +1,1 @@
+export { default } from '../lib/themes/jobsDbClassic';
